+++ src/main/scala/de/upb/cs/uc4/hyperledger/utilities/helper/Logger.scala
@@ -1,6 +1,11 @@
 package de.upb.cs.uc4.hyperledger.utilities.helper
 
+import org.slf4j.LoggerFactory
+import org.slf4j.{ Logger => Slf4jLogger }
+
 protected[hyperledger] object Logger {
+
+  private final val log: Slf4jLogger = LoggerFactory.getLogger("HyperledgerLogger")
 
   /** Logger utility to encapsulate printing error messages.
     * Maybe we can have some flag set to enable/disable in the future.
@@ -13,29 +18,30 @@ protected[hyperledger] object Logger {
               [MESSAGE] :: $message
               [ERROR] :: ${getInfoFromThrowable(exception)}
               """
-    println(msg)
+    log.error(msg)
     exception
+
   }
 
   /** Logger utility to encapsulate printing warning messages.
     * Maybe we can have some flag set to enable/disable in the future.
     */
   def warn(message: String): Unit = {
-    println(s"[WARNING] :: $message")
+    log.warn(message)
   }
 
   /** Logger utility to encapsulate printing debug messages.
     * Maybe we can have some flag set to enable/disable in the future.
     */
   def debug(message: String): Unit = {
-    println(s"[DEBUG] :: $message")
+    log.debug(message)
   }
 
   /** Logger utility to encapsulate printing info messages.
     * Maybe we can have some flag set to enable/disable in the future.
     */
   def info(message: String): Unit = {
-    println(s"[INFO] :: $message")
+    log.info(message)
   }
 
   def getInfoFromThrowable(exception: Throwable): String = {