+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# <a id="v0.17.3" />[v0.17.3](https://github.com/upb-uc4/hyperledger_api/compare/v0.17.2...v0.17.3) (2021-02-03)
+## Feature
+- add SLF4J Logger
+
 # <a id="v0.17.2" />[v0.17.2](https://github.com/upb-uc4/hyperledger_api/compare/v0.17.1...v0.17.2) (2021-02-02)
 
 ## Feature